updated QA

+++ mern-crash-course/back-end/controllers/product.controller.js
@@ -2,7 +2,7 @@ import Product from "../models/product.model.js";
 import mongoose from "mongoose";
 
 
-export const getProducts = async (req, res) => {
+export const updateProduct = async (req, res) => {
     const { id } = req.params;
     const product = req.body;
 
@@ -21,12 +21,15 @@ export const getProducts = async (req, res) => {
 export const deleteProduct = async (req, res) => {
     const { id } = req.params;
 
+    if(!mongoose.Types.ObjectId.isValid(id)){ 
+        return res.status(404).json({ success: false, message: "Invalid Product Id" });
+    }
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({ success: true, message: "Product deleted" });
     } catch (error) {
         console.log("Error in deleting products:", error.message);
-        res.status(404).json({ success: false, message: "Product not found" });
+        res.status(500).json({ success: false, message: "Server Error" });
     }
 };
